garage_checker/requirements.txt
# Garage Checker Requirements
numpy
opencv-python
torch
torchvision
ultralytics
fast-plate-ocr
python-dotenv
python-telegram-bot
PyYAML
